requirements.txt
reportlab
python-docx
sympy
jinja2
pathlib
google-adk
flask
python-dotenv